+++ src/Function.OpenApi/OpenApiRequestBodyAttribute.cs
@@ -1,5 +1,6 @@
 ﻿namespace Function.OpenApi;
 
+[AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
 public class OpenApiRequestBodyAttribute(string contentType, Type bodyType) : Attribute
 {
     public string ContentType => contentType;